Make sure element from xpath loaded before execute
Before extract/fill/click to element, wait until it load completed
before execute it, so you don't need to use sleep function to wait it.
However when the element loaded but it not show in the GUI we need to
use the sleep function to wait until the element show.